+++ src/main/java/dev/shadowsoffire/attributeslib/mixin/PortLibStepHeightFix.java
@@ -12,12 +12,16 @@ public class PortLibStepHeightFix {
     // TODO remove, needed as I can't update porting lib to latest version without moving to entity refactor branch
 
     @ModifyReturnValue(method = "createLivingAttributes", at = @At("RETURN"))
-    private static AttributeSupplier.Builder zenith_attributes_addPortLibAttribute(AttributeSupplier.Builder builder) {
+    private static AttributeSupplier.Builder zenith_attributes$addPortLibAttribute(AttributeSupplier.Builder builder) {
         return builder.add(PortingLibAttributes.STEP_HEIGHT_ADDITION);
     }
 
     @ModifyReturnValue(method = "maxUpStep", at = @At("RETURN"))
-    private float zenith_attributes_modifyStepHeight(float vanillaStep) {
-        return (float) (vanillaStep + ((LivingEntity) (Object) this).getAttribute(PortingLibAttributes.STEP_HEIGHT_ADDITION).getValue());
+    private float zenith_attributes$modifyStepHeight(float vanillaStep) {
+        var portLibStepHeight = ((LivingEntity) (Object) this).getAttribute(PortingLibAttributes.STEP_HEIGHT_ADDITION);
+        if (portLibStepHeight != null) {
+            return (float) Math.max(0, (vanillaStep + portLibStepHeight.getValue()));
+        }
+       return vanillaStep;
     }
 }