+++ changelog.md
@@ -1,7 +1,11 @@
 
 
+## 0.1.2
+* Added trinkets support, so trinkets will show in the attributes screen
+* 
+
 ## 0.1.1
-* Update the config to 
+* Update the config to be less horrible
 
 ## 0.1.0
 * Fix step height attribute being broken